Modifica int a long

+++ JAVA/NewDMApp/src/main/java/db/ElencaDao.java
@@ -31,12 +31,12 @@ public class ElencaDao {
   }
 
   public static void checkCorrispondenza(
-      int codiceScontrino, String dataScontrino, long codiceProdotto)
+      long codiceScontrino, String dataScontrino, long codiceProdotto)
       throws DatabaseException, ElencaException {
 
     try {
       PreparedStatement state = DatabaseConnection.con.prepareStatement(Query.elencaCheck);
-      state.setInt(1, codiceScontrino);
+      state.setLong(1, codiceScontrino);
       state.setString(2, dataScontrino);
       state.setLong(3, codiceProdotto);
       ResultSet rs = state.executeQuery();
